Moved ConfNode out into its own module definition
